Crate repitile_core

repitile_core contains the core functionality for monitoring and regulating the environment of the reptile tank.

See individual struct/module documentation for better explanations of each part.

Modules

config
Contains Configuration related functions and types.

profile
Contains Profile related functions and types.

regulator
Regulator trait.

sensor
Sensor trait.

Structs

CURRENT_CONDITIONS
Global current conditions that can be accessed by other threads.

Core
The core structure that monitors the environment.

CurrentConditions
Describes the current conditions of the environment.

Enums

CommReq
Describes messages passed between the core and server.